slice: make u16 index buffers

+++ src/slice.rs
@@ -29,12 +29,12 @@ pub fn slice_for_fan<R, F>(factory: &mut F, size: u32) -> gfx::Slice<R>
     where R: gfx::Resources,
           F: gfx::Factory<R>,
 {
-    let mut indices: Vec<u32> = Vec::with_capacity((size as usize) * 2 - 2);
+    let mut indices: Vec<u16> = Vec::with_capacity((size as usize) * 2 - 2);
 
     for (i, j) in (1..size).zip(2..size) {
         indices.push(0);
-        indices.push(i);
-        indices.push(j);
+        indices.push(i as u16);
+        indices.push(j as u16);
     }
 
     gfx::Slice {
